memory-service/requirements.txt
fastapi>=0.116,<1
uvicorn>=0.35,<1
mem0ai>=1,<2
chromadb>=1,<2
fastembed>=0.7,<1
